+++ DirectX2D/Framework/Utilities/BinaryWriter.cpp
@@ -1,5 +1,4 @@
 #include "Framework.h"
-#include "BinaryWriter.h"
 
 BinaryWriter::BinaryWriter(string filePath)
 {
@@ -37,15 +36,15 @@ void BinaryWriter::String(string data)
 	UInt((UINT)data.size());
 
 	const char* str = data.c_str();
-	WriteFile(file, str, sizeof(char)*data.size(), &size, nullptr);
+	WriteFile(file, str, sizeof(char)*(DWORD)data.size(), &size, nullptr);
 }
 
 void BinaryWriter::WString(wstring data)
 {
-	UInt(data.size());
+	UInt((UINT)data.size());
 
 	const WCHAR* str = data.c_str();
-	WriteFile(file, str, sizeof(WCHAR) * data.size(), &size, nullptr);
+	WriteFile(file, str, sizeof(WCHAR) * (DWORD)data.size(), &size, nullptr);
 }
 
 void BinaryWriter::Byte(void* data, UINT dataSize)
